requirements.txt
google-generativeai==0.5.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
streamlit==1.35.0
python-dotenv==1.0.1
google-cloud-aiplatform==1.44.0
google-auth==2.30.0
google-api-python-client==2.125.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.32.3